+++ R/highest_concentration.R
@@ -1,6 +1,8 @@
-#' Highest concentration
+#' Highest concentration risk
 #'
-#' @description Find the centre coordinates of a circle with a fixed radius that maximizes the coverage of total fire risk insured.
+#' @description Find the centre coordinates of a circle with a fixed radius that maximizes the coverage of total fire risk insured. `highest_concentration()`
+#' returns the coordinates (lon/lat) and the corresponding concentration. The concentration is defined as the sum of all observations
+#' within a circle of a certain radius. See \code{\link{concentration}} for determining concentration for pre-defined coordinates.
 #'
 #' @param df data.frame of locations, should at least include column for longitude, latitude and sum insured
 #' @param value column name with value of interest to summarize (e.g. sum insured)
@@ -11,48 +13,123 @@
 #' @param grid_distance distance (in meters) for precision of concentration risk (default is 20m).
 #'     `neighborhood_search()` can be used to search for coordinates with even higher concentrations
 #'     in the neighborhood of the highest concentrations.
-#' @param gh_precision positive integer to define geohash precision. The precision controls the 'zoom level'.
-#'     5 is 4.89 x 4.89km; 6 is 1.22km x 0.61km; 7 is 153m x 153m; 8 is 39m x 19m. Defaults to 6. Precision level
-#'     of 6 must be used for a radius of 200m.
+#' @param gh_precision positive integer to define geohash precision. See details.
+#' @param display_progress show progress bar (TRUE/FALSE). Defaults to TRUE.
 #'
-#' @import data.table
-#' @import geohashTools
+#' @details A recently European Commission regulation requires insurance
+#' companies to determine the maximum value of insured fire risk policies of
+#' all buildings that are partly or fully located within circle of a radius of 200m
+#' (Commission Delegated Regulation (EU), 2015, Article 132). The problem can
+#' be stated as: "find the centre coordinates of a circle with a fixed radius
+#' that maximizes the coverage of total fire risk insured". This can be viewed
+#' as a particular instance of the Maximal Covering Location Problem (MCLP)
+#' with fixed radius. See Gomes (2018) for a solution to the maximum fire
+#' risk insured capital problem using a multi-start local search meta-heuristic.
+#' The computational performance of \code{highest_concentration()} is
+#' investigated to overcome the long times the MCLP algorithm is taking.
+#' \code{highest_concentration()} is written in C++, and for 500,000 buildings
+#' it needs about 5-10 seconds to determine the maximum value of insured fire risk
+#' policies that are partly or fully located within circle of a radius of 200m.
+#'
+#' `highest_concentration()` uses Gustavo Niemeyer's wonderful and elegant
+#' geohash coordinate system. Niemeyer's Geohash method encodes latitude and
+#' longitude as binary string where each binary value derived from a decision
+#' as to where the point lies in a bisected region of latitude or longitudinal
+#' space. The first step is to convert all latitude/longitude coordinates into
+#' geohash-encoded strings.
+#'
+#' The length of the geohash (`gh_precision`) controls the 'zoom level':
+#' \itemize{
+#' \item precision 5 is 4.89 x 4.89km;
+#' \item precision 6 is 1.22km x 0.61km;
+#' \item precision 7 is 153m x 153m;
+#' \item precision 8 is 39m x 19m.
+#' }
+#'
+#' For a circle with a radius of 200m the precision of the geohash should
+#' be set equal to 6 (default). Then the `value` column is aggregated (sum)
+#' per geohash (with a buffer of size `radius` around each geohash,
+#' since the coordinates of the highest concentration can be near the edge of
+#' the geohash). The geohashes with a aggregated value below the lowerbound
+#' are removed, where the lowerbound is equal to the maximum of the
+#' `value` column. Then a grid is created, with a distance of `grid_distance`
+#' between the points. See example section for a illustration of the algorithm.
+#' As a last step for each grid point the concentration is calculated.
+#'
+#' @importFrom data.table data.table
+#' @importFrom data.table as.data.table
+#' @importFrom data.table setcolorder
+#' @importFrom geohashTools gh_encode
+#' @importFrom geohashTools gh_decode
 #'
 #' @author Martin Haringa
 #'
-#' @return data.frame
+#' @return data.frame with coordinates (lon/lat) with the highest concentrations
+#'
+#' @references Commission Delegated Regulation (EU) (2015). Solvency II Delegated Act 2015/35. Official Journal of the European Union, 58:124.
+#' @references Gomes M.I., Afonso L.B., Chibeles-Martins N., Fradinho J.M. (2018). Multi-start Local Search Procedure for the Maximum Fire Risk Insured Capital Problem. In: Lee J., Rinaldi G., Mahjoub A. (eds) Combinatorial Optimization. ISCO 2018. Lecture Notes in Computer Science, vol 10856. Springer, Cham. <doi:10.1007/978-3-319-96151-4_19>
+#'
 #' @examples
 #'  \dontrun{
 #' # Find highest concentration with a precision of a grid of 20 meters
 #' hc1 <- highest_concentration(Groningen, amount, radius = 200, grid_distance = 20)
 #'
-#' # Set distance between grid points to 10 meters
-#' hc2 <- highest_concentration(Groningen, amount, radius = 200, grid_distance = 10)
-#'
-#' # Increase the number of calls to the concentration function for more extensive search
+#' # Look for coordinates with even higher concentrations in the
+#' # neighborhood of the coordinates with the highest concentration
 #' hc1_nghb <- neighborhood_gh_search(hc1, max.call = 7000)
-#' hc2_nghb <- neighborhood_gh_search(hc2, max.call = 7000)
 #' print(hc1_nghb)
-#' print(hc2_nghb)
+#'
+#' # Create map with geohashes above the lowerbound
+#' # The highest concentration lies in one of the geohashes
+#' plot(hc1)
+#'
+#' # Create map with highest concentration
+#' plot(hc1_nghb)
 #' }
 #'
 #' @export
-highest_concentration <- function(df, value, lon = lon, lat = lat, lowerbound = NULL, radius = 200, grid_distance = 20,
-                                  gh_precision = 6, display_progress = TRUE){
+highest_concentration <- function(df, value, lon = lon, lat = lat,
+                                  lowerbound = NULL, radius = 200,
+                                  grid_distance = 20, gh_precision = 6,
+                                  display_progress = TRUE){
 
   value_nm <- deparse(substitute(value))
   lon_nm <- deparse(substitute(lon))
   lat_nm <- deparse(substitute(lat))
+  df_nm <- deparse(substitute(df))
+
+  if ( !all(c(value_nm) %in% names(df))) {
+    stop(df_nm, " does not contain column specified in `value`. Specify with argument `value`.",
+         call. = FALSE)
+  }
+
+  if ( !all(c(lon_nm, lat_nm) %in% names(df))) {
+    stop(df_nm, " does not contain columns ", lon_nm, " and ", lat_nm,
+         ". Specify with arguments `lon` and `lat`.", call. = FALSE)
+  }
+
+  nrows1 <- nrow(df)
+
+  # Remove rows with NA values
+  df <- df[!is.na(df[[value_nm]]), ]
+
+  if ( nrows1 !=  nrow(df) ){
+    warning(nrows1 - nrow(df), " NAs detected in ", value_nm, ". Rows with NAs removed.", call. = FALSE)
+  }
 
   # Add geohash (length 5 is 4.89km x 4.89km; length 6 is 1.22 x 0.61km; length 7 is 153m x 153m; length 8 is 38m x 19m)
   df[["geohash"]] <- geohashTools::gh_encode(latitude = df[[lat_nm]],
-                                          longitude = df[[lon_nm]],
-                                          precision = gh_precision)
+                                             longitude = df[[lon_nm]],
+                                             precision = gh_precision)
 
   # Determine lower bound for concentration (based on circles around 1000 highest sums insured)
   # This only works with some relatively high sums insured
   if ( is.null(lowerbound) ){
-    lowerbound_1 <- lower_bound_fn(df, value_nm, radius = radius, highest = 1000)
+    lowerbound_1 <- lower_bound_fn(df, value_nm,
+                                   lat_nm = lat_nm,
+                                   lon_nm = lon_nm,
+                                   radius = radius,
+                                   highest = 1000)
   }
 
   portfolio_dt <- data.table::data.table(df)
@@ -64,7 +141,12 @@ highest_concentration <- function(df, value, lon = lon, lat = lat, lowerbound = 
   gh_sum_nghbrs <- add_gh_nghbrs_sum(gh_sum, "gh_self_sum")
 
   # Find lowerbound for use cases where sum insured is same for all coords
-  lowerbound_2 <- lower_bound_fn2(gh_sum_nghbrs, portfolio_dt, value_nm, radius = radius)
+  lowerbound_2 <- lower_bound_fn2(gh_sum_nghbrs,
+                                  full = portfolio_dt,
+                                  col = value_nm,
+                                  lat_nm = lat_nm,
+                                  lon_nm = lon_nm,
+                                  radius = radius)
 
   # Take max of both lowerbounds
   lowerbound <- max(lowerbound_1, lowerbound_2)
@@ -83,7 +165,11 @@ highest_concentration <- function(df, value, lon = lon, lat = lat, lowerbound = 
 
   # Determine sum of remaining points within bounding box of radius
   # around remaining geohashes
-  bbox_sum <- add_gh_bbox_sum(gh_center, pts_remaining, radius = radius)
+  bbox_sum <- add_gh_bbox_sum(gh_center,
+                              pts_remaining,
+                              lon_nm = lon_nm,
+                              lat_nm = lat_nm,
+                              radius = radius)
   gh_sum_bbox <- cbind(gh_remaining, bbox_sum)
 
   # Remove (again) geohashes with a total sum insured lower than the lower bound
@@ -96,37 +182,44 @@ highest_concentration <- function(df, value, lon = lon, lat = lat, lowerbound = 
   gh_grid <- create_grid_points(gh_remaining_bbox, meters = grid_distance)
 
   # Concentration for each grid point
-  gh_grid_conc <- concentration(gh_grid, pts_remaining, `_sum_insured`,
+  names(pts_remaining)[names(pts_remaining) == lat_nm] <- "lat"
+  names(pts_remaining)[names(pts_remaining) == lon_nm] <- "lon"
+  names(pts_remaining2)[names(pts_remaining2) == lat_nm] <- "lat"
+  names(pts_remaining2)[names(pts_remaining2) == lon_nm] <- "lon"
+
+  gh_grid_conc <- concentration(gh_grid,
+                                pts_remaining,
+                                value = `_sum_insured`,
                                 radius = radius,
                                 display_progress = display_progress)
 
   gh_grid_conc_sort <- gh_grid_conc[order(-concentration)]
+  data.table::setcolorder(gh_grid_conc_sort, c("concentration", "lon", "lat", "geohash"))
 
   attr(gh_grid_conc_sort, "value_nm") <- value_nm
   attr(gh_grid_conc_sort, "pts_remaining") <- pts_remaining2
   attr(gh_grid_conc_sort, "gh_remaining") <- gh_remaining_bbox
   attr(gh_grid_conc_sort, "radius") <- radius
 
-  class(gh_grid_conc_sort) <- append(class(gh_grid_conc_sort), "concentration")
+  class(gh_grid_conc_sort) <- append("concentration", class(gh_grid_conc_sort))
   return(gh_grid_conc_sort)
 }
 
 
 #' Search for coordinates with higher concentrations within geohash
 #'
-#' `highest_concentration()` returns the highest concentration within a portfolio based on a grid.
-#' The distance (in meters) within the points in the grid can be set by the `grid_distance` argument.
+#' \code{\link{highest_concentration}} returns the highest concentration within a portfolio based on grid points.
 #' However, higher concentrations can be found within two grid points. `neighborhood_gh_search()`
 #' looks for even higher concentrations in the neighborhood of the grid points with the highest
 #' concentrations. This optimization is done by means of Simulated Annealing.
 #'
 #' @param hc object of class `concentration` obtained from `highest_concentration()`
-#' @param geohash character vector with the name of the geohash (defaults to NULL)
 #' @param highest_geohash the number of geohashes the searching algorithm is applied to.
 #'     Defaults to 1 (i.e. algorithm is only applied to the geohash with the highest concentration).
 #' @param max.call maximum number of calls to the concentration function
 #'     (i.e. the number of coordinates in the neighborhood of the highest concentration). Defaults to 1000.
 #' @param verbose show messages from the algorithm (TRUE/FALSE). Defaults to FALSE.
+#' @param seed set seed
 #'
 #' @importFrom GenSA GenSA
 #' @importFrom geohashTools gh_decode
@@ -134,6 +227,7 @@ highest_concentration <- function(df, value, lon = lon, lat = lat, lowerbound = 
 #' @author Martin Haringa
 #'
 #' @return data.frame
+#'
 #' @examples
 #' \dontrun{
 #' # Find highest concentration with a precision of a grid of 20 meters
@@ -146,11 +240,11 @@ highest_concentration <- function(df, value, lon = lon, lat = lat, lowerbound = 
 #' hc1_nghb <- neighborhood_gh_search(hc1, max.call = 7000)
 #' hc2_nghb <- neighborhood_gh_search(hc2, max.call = 7000)
 #' print(hc1_nghb)
-#' print(hc2_nghb)
+#' plot(hc1_nghb)
 #' }
 #'
 #' @export
-neighborhood_gh_search <- function(hc, highest_geohash = 1, max.call = 1000, verbose = TRUE){
+neighborhood_gh_search <- function(hc, highest_geohash = 1, max.call = 1000, verbose = TRUE, seed = 1){
 
   if( !inherits(hc, c("concentration")) ) {
     stop("Input must be of class concentration. Use highest_concentration() first.", call. = FALSE)
@@ -186,6 +280,7 @@ neighborhood_gh_search <- function(hc, highest_geohash = 1, max.call = 1000, ver
     upper <- c(lon_end, lat_end)
 
     # Generate coordinates
+    set.seed(seed)
     sa <- GenSA::GenSA(lower = lower, upper = upper,
                        fn = csa,
                        control = list(verbose = verbose,
@@ -201,29 +296,63 @@ neighborhood_gh_search <- function(hc, highest_geohash = 1, max.call = 1000, ver
   dc <- do.call(rbind, gsa_lst)
   attr(dc, "pts_remaining") <- pts_remaining
   attr(dc, "radius") <- radius
+  attr(dc, "value_nm") <- value_nm
+
+  class(dc) <- append("neighborhood", class(dc))
   return(dc)
 }
 
 
-#' Test 2
+#' Automatically create a plot for objects obtained from highest_concentration()
 #'
-#' @param hc highest concentration
+#' @description Takes an object produced by `highest_concentration()`, and creates an interactive map.
 #'
-#' @return sdf
+#' @param x object of class `concentration` obtained from `highest_concentration()`
+#' @param grid_points show grid points (TRUE), or objects (FALSE)
+#' @param legend_title title of legend
+#' @param ... additional arguments affecting the interactive map produced
+#'
+#' @return Interactive view of geohashes with highest concentrations
+#'
 #' @author Martin Haringa
 #'
-#' @import data.table
 #' @importFrom dplyr group_by
 #' @importFrom dplyr summarise
-#' @import sf
-#' @import mapview
+#' @importFrom sf st_as_sf
+#' @importFrom sf st_cast
+#' @importFrom mapview mapview
 #'
 #' @export
-f1 <- function(hc){
+plot.concentration <- function(x, grid_points = TRUE, legend_title = NULL, ...){
 
-  pts <- attr(hc, "pts_remaining")
-  geohashes0 <- attr(hc, "gh_remaining")
-  pts_sf <- sf::st_as_sf(pts, coords = c("lon", "lat"), crs = 4326)
+  if (!inherits(x, "concentration")) {
+    stop("plot.concentration requires an object of class concentration", call. = FALSE)
+  }
+
+  pts <- attr(x, "pts_remaining")
+  geohashes0 <- attr(x, "gh_remaining")
+  value_nm <- attr(x, "value_nm")
+  legend_nm <- value_nm
+  x_nm <- deparse(substitute(x))
+
+  if ( !is.null(legend_title) ){
+    legend_nm <- legend_title
+  }
+
+  if ( !isTRUE(grid_points) ){
+    pts_sf <- sf::st_as_sf(pts, coords = c("lon", "lat"), crs = 4326)
+  }
+
+  if ( isTRUE(grid_points) ){
+
+    if ( !all(c("lon", "lat", "concentration") %in% names(x))) {
+      stop(x_nm, " must contain columns `concentration`, `lon` and `lat`", call. = FALSE)
+    }
+
+    legend_nm <- "Concentration"
+    value_nm <- "concentration"
+    pts_sf <- sf::st_as_sf(x, coords = c("lon", "lat"), crs = 4326)
+  }
 
   geohashes <- geohashes0[
     , north := latitude + delta_latitude][
@@ -246,45 +375,75 @@ f1 <- function(hc){
 
   mapview::mapview(rect_sf, col.regions = "skyblue", legend = FALSE,
                    alpha.regions = .2) +
-    mapview::mapview(pts_sf, zcol = "amount")
+    mapview::mapview(pts_sf, zcol = value_nm, layer.name = legend_nm)
 }
 
-#' Test1
+
+
+#' Automatically create a plot for objects obtained from neighborhood_gh_search()
 #'
-#' @param hc object
+#' @description Takes an object produced by `neighborhood_gh_search()`, and creates an interactive map.
 #'
-#' @return Interactive view of highest concentration on top of base map
+#' @param x object neighborhood object produced by `neighborhood_gh_search()`
+#' @param buffer numeric value, show objects within buffer (in meters) from circle (defaults to 0)
+#' @param legend_title title of legend
+#' @param ... additional arguments affecting the interactive map produced
+#'
+#' @return Interactive view of highest concentration on map
 #'
 #' @author Martin Haringa
 #'
-#' @import sf
-#' @import mapview
+#' @importFrom sf st_as_sf
+#' @importFrom sf st_transform
+#' @importFrom sf st_buffer
+#' @importFrom sf st_geometry
+#' @importFrom mapview mapview
+#' @importFrom RColorBrewer brewer.pal
+#'
+#' @rdname plot
 #'
 #' @export
-f2 <- function(hc){
-  pts0 <- attr(hc, "pts_remaining")
-  radius <- attr(hc, "radius")
+plot.neighborhood <- function(x, buffer = 0, legend_title = NULL, ...){
 
-  pts_lst <- vector("list", nrow(hc))
-  for ( i in 1:nrow(hc)){
+  if (!inherits(x, "neighborhood")) {
+    stop("plot.neighborhood requires an object of class neighborhood", call. = FALSE)
+  }
+
+  pts0 <- attr(x, "pts_remaining")
+  radius <- attr(x, "radius")
+  value_nm <- attr(x, "value_nm")
+  legend_nm <- value_nm
+
+  if ( !is.null(legend_title) ){
+    legend_nm <- legend_title
+  }
+
+  pts_lst <- vector("list", nrow(x))
+  for ( i in 1:nrow(x)){
     pts_lst[[i]] <- points_in_circle(pts0,
-                                     lon_center = hc$lon[i],
-                                     lat_center = hc$lat[i],
-                                     radius = radius + 200)
+                                     lon_center = x$lon[i],
+                                     lat_center = x$lat[i],
+                                     radius = radius + buffer)
   }
 
   pts <- do.call(rbind, pts_lst)
   pts_sf <- sf::st_as_sf(pts, coords = c("lon", "lat"), crs = 4326)
 
-  circle_sf <- hc %>%
+  circle_sf <- x %>%
     sf::st_as_sf(coords = c("lon", "lat"), crs = 4326) %>%
     sf::st_transform(3035) %>%
     sf::st_buffer(dist = units::set_units(radius, "meters")) %>%
-    sf::st_transform(4326) %>%
-    sf::st_geometry()
+    sf::st_transform(4326) #%>%
+  #  sf::st_geometry()
 
-  mapview::mapview(pts_sf, zcol = "amount") +
-    mapview::mapview(circle_sf, col.regions = "red")
+  suppressWarnings({
+    mapview::mapview(circle_sf,
+                     col.regions = rev(RColorBrewer::brewer.pal(nrow(circle_sf), "Reds")),
+                     zcol = "highest_concentration",
+                     legend = TRUE,
+                     layer.name = "Highest concentration") +
+      mapview::mapview(pts_sf, zcol = value_nm, layer.name = legend_nm)
+  })
 }