+++ R/utils.R
@@ -5,16 +5,17 @@
 #' @importFrom utils head
 #'
 #' @keywords internal
-lower_bound_fn <- function(df, col, radius = 200, highest = 50){
-
-  if ( !all(c("lon", "lat") %in% names(df))) {
-    stop("columns lat and lon do not exist in df", call. = FALSE)
-  }
+lower_bound_fn <- function(df, col, lat_nm, lon_nm, radius = 200, highest = 50){
 
   df_order <- df[order(-df[[col]]),]
   high_df <- head(df_order, highest)
-  high_df_pts <- high_df[,c("lon","lat")]
+  high_df_pts <- high_df[,c(lon_nm, lat_nm)]
+  colnames(high_df_pts) <- c("lon", "lat")
+
   names(df)[names(df) == col] <- "valueconc"
+  names(df)[names(df) == lat_nm] <- "lat"
+  names(df)[names(df) == lon_nm] <- "lon"
+
   conc <- concentration(high_df_pts, df, valueconc,
                         display_progress = FALSE,
                         radius = radius)
@@ -22,9 +23,11 @@ lower_bound_fn <- function(df, col, radius = 200, highest = 50){
 }
 
 #' @importFrom utils head
+#' @importFrom data.table setorder
 #'
 #' @keywords internal
-lower_bound_fn2 <- function(df, full, col, radius = 200, highest = 100){
+lower_bound_fn2 <- function(df, full, col, lat_nm, lon_nm,
+                            radius = 200, highest = 100){
 
   df_order <- data.table::setorder(df, -geohash_sum)
   high_df <- head(df_order, highest)
@@ -32,6 +35,8 @@ lower_bound_fn2 <- function(df, full, col, radius = 200, highest = 100){
   high_df_gh <- data.frame(high_df_gh_lst)
   names(high_df_gh) <- c("lat", "lon")
   names(full)[names(full) == col] <- "valueconc"
+  names(full)[names(full) == lat_nm] <- "lat"
+  names(full)[names(full) == lon_nm] <- "lon"
   conc <- concentration(high_df_gh, full, valueconc,
                         display_progress = FALSE,
                         radius = radius)
@@ -41,14 +46,13 @@ lower_bound_fn2 <- function(df, full, col, radius = 200, highest = 100){
 #' Determine total insured sum of all neighbours (including hash itself)
 #' df should contain a column called hash and a column with sum per hash
 #'
-#' @import data.table
+#' @importFrom data.table setDT
+#' @importFrom data.table setnames
+#' @importFrom data.table melt
+#' @importFrom data.table setkey
 #'
 #' @keywords internal
 add_gh_nghbrs_sum <- function(df, colname_sum){
-
-  if(!"geohash" %in% colnames(df)){
-    stop("df should contain column with hash", call. = FALSE)
-  }
 
   x_dt <- data.table::setDT(df, key = "geohash")
   data.table::setnames(x_dt, old = colname_sum, new = "geohash_sum")
@@ -74,6 +78,7 @@ add_gh_nghbrs_sum <- function(df, colname_sum){
 }
 
 
+#' Determine total sum for each geohash with buffer of 200m
 #'
 #' @param hash_decode data.frame as output from `geohashTools::gh_decode()`
 #' @param full data.frame with portfolio with lon, lat and sum insured
@@ -83,15 +88,15 @@ add_gh_nghbrs_sum <- function(df, colname_sum){
 #' @importFrom Rcpp evalCpp
 #'
 #' @keywords internal
-add_gh_bbox_sum <- function(hash_decode, full, radius){
+add_gh_bbox_sum <- function(hash_decode, full, lon_nm, lat_nm, radius){
 
   sub_df <- data.frame("lon" = hash_decode[["longitude"]],
                        "lat" = hash_decode[["latitude"]],
                        "delta_latitude" = hash_decode[["delta_latitude"]],
                        "delta_longitude" = hash_decode[["delta_longitude"]])
 
-  full_df <- data.frame("lon" = full[["lon"]],
-                        "lat" = full[["lat"]],
+  full_df <- data.frame("lon" = full[[lon_nm]],
+                        "lat" = full[[lat_nm]],
                         "value" = full[["_sum_insured"]])
 
   block_df <- block_loop_cpp(sub_df, full_df, radius, display_progress = FALSE)
@@ -101,6 +106,10 @@ add_gh_bbox_sum <- function(hash_decode, full, radius){
 }
 
 
+
+#'
+#' @importFrom data.table rbindlist
+#'
 #' @keywords internal
 create_grid_points <- function(df, meters, hash = geohash,
                                lon = longitude, lat = latitude,
@@ -133,6 +142,6 @@ create_grid_points <- function(df, meters, hash = geohash,
 }
 
 
-#' @importFrom magrittr %>%
+#' @importFrom dplyr %>%
 #' @export
-magrittr::`%>%`
+dplyr::`%>%`